Split knit bundle.rs: extract the prune subsystem
Move the dead-work prune logic (assessment, caching, orphan worktrees, and
KnitHub remote orphan detection) out of src/commands/bundle.rs into
bundle/prune.rs. bundle/mod.rs keeps bundle lifecycle/inspection (show, list,
validate, switch, archive, restore, delete) and the shared bundle-state and
validation helpers. No behavior change; tests and clippy unchanged.


Knit-Group: kg_20260530_5da816
Knit-Bundle: split-oversized-source-files
